Refactoring + Working Progress on Workstation

+++ Assets/Scripts/Objects Scripts/Workstation.cs
@@ -4,6 +4,7 @@ public class Workstation : InteractableAsset
 {
 
     [SerializeField] private SceneObjectTransformSO[] sceneObjectTransformArray;
+    private int workProgress;
 
     // Function that defines the logic when player interacts with the Counter.
     public override void Interact(Player player)
@@ -16,6 +17,7 @@ public class Workstation : InteractableAsset
                 if(HasObjectTransform(player.GetSceneObject().GetSceneObjectSO()))
                 {
                     player.GetSceneObject().SetSceneObjectParent(this);
+                    workProgress = 0;
                 }
             }
         }
@@ -32,36 +34,46 @@ public class Workstation : InteractableAsset
     {
         if(HasSceneObject() && HasObjectTransform(GetSceneObject().GetSceneObjectSO()))
         {
-            SceneObjectSO outputSceneObjectSO = GetOutputForInput(GetSceneObject().GetSceneObjectSO());
-            GetSceneObject().DeleteObject();
+            workProgress++;
+            SceneObjectTransformSO sceneObjectTransformSO = GetSceneObjectTransformSO(GetSceneObject().GetSceneObjectSO());
+            if(workProgress >= sceneObjectTransformSO.workProgressMax)
+            {
+                SceneObjectSO outputSceneObjectSO = GetOutputForInput(GetSceneObject().GetSceneObjectSO());
+                GetSceneObject().DeleteObject();
 
-            SceneObject.SpawnSceneObject(outputSceneObjectSO, this);
-            Debug.Log("Player/Workstation Interaction - Deleting Object and Creating New One");
+                SceneObject.SpawnSceneObject(outputSceneObjectSO, this);
+                Debug.Log("Player/Workstation Interaction - Deleting Object and Creating New One");
+            }
         }
     }
 
     private bool HasObjectTransform(SceneObjectSO inputSceneObjectSO)
     {
-        foreach(SceneObjectTransformSO sceneObjectTransformSO in sceneObjectTransformArray)
-        {
-            if(sceneObjectTransformSO.input == inputSceneObjectSO)
-            {
-                return true;
-            }
-        }
-        return false;
-        
+        SceneObjectTransformSO sceneObjectTransformSO = GetSceneObjectTransformSO(inputSceneObjectSO);
+        return sceneObjectTransformSO != null;
     }
 
     private SceneObjectSO GetOutputForInput(SceneObjectSO inputSceneObjectSO)
+    {
+
+        SceneObjectTransformSO sceneObjectTransformSO = GetSceneObjectTransformSO(inputSceneObjectSO);
+        if(sceneObjectTransformSO != null)
+        {
+            return sceneObjectTransformSO.output;
+        }
+        return null;
+    }
+
+    private SceneObjectTransformSO GetSceneObjectTransformSO(SceneObjectSO inputSceneObjectSO)
     {
         foreach(SceneObjectTransformSO sceneObjectTransformSO in sceneObjectTransformArray)
         {
             if(sceneObjectTransformSO.input == inputSceneObjectSO)
             {
-                return sceneObjectTransformSO.output;
+                return sceneObjectTransformSO;
             }
         }
         return null;
+        
     }
 }
